fix:major update in component rendering
