+++ src/components/layouts/EmptyState/DashboardEmptyState.tsx
@@ -1,34 +1,15 @@
 'use client'
 
-import { handleCreate } from '@/app/actions'
-import { NewTaskForm } from '@/app/ui/NewTaskForm'
 import { PrimaryBtn } from '@/components/buttons/PrimaryBtn'
 import { AppMargin, SizeofAppMargin } from '@/hoc/AppMargin'
 import { AddIcon, TasksListIcon } from '@/icons'
-import { clearCreateTaskFields, selectCreateTask, setShowModal } from '@/redux/features/createTaskSlice'
-import { appendTask, selectTaskBoard } from '@/redux/features/taskBoardSlice'
+import { setShowModal } from '@/redux/features/createTaskSlice'
 import store from '@/redux/store'
-import { CreateAttachmentRequest } from '@/types/dto/attachments.dto'
-import { CreateTaskRequestSchema } from '@/types/dto/tasks.dto'
-import { ISignedUrlUpload, UserType } from '@/types/interfaces'
-import { bulkRemoveAttachments } from '@/utils/bulkRemoveAttachments'
+import { UserType } from '@/types/interfaces'
 import { SxCenter } from '@/utils/mui'
-import { Box, Modal, Stack, Typography } from '@mui/material'
-import { useSelector } from 'react-redux'
+import { Box, Stack, Typography } from '@mui/material'
 
-const DashboardEmptyState = ({
-  getSignedUrlUpload,
-  handleCreateMultipleAttachments,
-  userType,
-}: {
-  getSignedUrlUpload?: (fileName: string) => Promise<ISignedUrlUpload>
-  handleCreateMultipleAttachments?: (attachments: CreateAttachmentRequest[]) => Promise<void>
-  userType: UserType
-}) => {
-  const { token } = useSelector(selectTaskBoard)
-  const { title, description, workflowStateId, assigneeId, assigneeType, attachments, dueDate, showModal } =
-    useSelector(selectCreateTask)
-
+const DashboardEmptyState = ({ userType }: { userType: UserType }) => {
   return (
     <>
       <AppMargin size={SizeofAppMargin.LARGE} py="20px">
@@ -77,49 +58,6 @@ const DashboardEmptyState = ({
           </Stack>
         </Box>
       </AppMargin>
-
-      {getSignedUrlUpload && handleCreateMultipleAttachments && (
-        <Modal
-          open={showModal}
-          onClose={async () => {
-            store.dispatch(setShowModal())
-            store.dispatch(clearCreateTaskFields())
-            await bulkRemoveAttachments(attachments)
-          }}
-          aria-labelledby="create-task-modal"
-          aria-describedby="add-new-task"
-        >
-          <NewTaskForm
-            handleCreate={async () => {
-              if (title) {
-                store.dispatch(setShowModal())
-                store.dispatch(clearCreateTaskFields())
-                const createdTask = await handleCreate(
-                  token as string,
-                  CreateTaskRequestSchema.parse({
-                    title,
-                    body: description,
-                    workflowStateId,
-                    assigneeType,
-                    assigneeId,
-                    dueDate,
-                  }),
-                )
-                store.dispatch(appendTask(createdTask))
-                const toUploadAttachments: CreateAttachmentRequest[] = attachments.map((el) => {
-                  return {
-                    ...el,
-                    taskId: createdTask.id,
-                  }
-                })
-                store.dispatch(clearCreateTaskFields())
-                await handleCreateMultipleAttachments(toUploadAttachments)
-              }
-            }}
-            getSignedUrlUpload={getSignedUrlUpload}
-          />
-        </Modal>
-      )}
     </>
   )
 }
